+++ Sources/TGReduxKit/Core/CancellationID.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+public struct CancellationID: Hashable, Sendable, RawRepresentable, ExpressibleByStringLiteral {
+    public let rawValue: String
+
+    public init(rawValue: String) {
+        self.rawValue = rawValue
+    }
+
+    public init(_ rawValue: String) {
+        self.rawValue = rawValue
+    }
+
+    public init(stringLiteral value: StringLiteralType) {
+        self.rawValue = value
+    }
+}